feat(condo): INFRA-517 added script for generate client utils
