+++ apps/condo/package.json
@@ -17,10 +17,8 @@
     "migrate": "./../../bin/kmigrator.py migrate",
     "migrate:down": "./../../bin/kmigrator.py down",
     "migrate:unlock": "./../../bin/kmigrator.py unlock",
-    "maketypes:legacy": "graphql-codegen -c codegen.yaml",
-    "maketypes:nextgen": "graphql-codegen -c codegen.ts",
-    "maketypes:schema": "mkdirp dist/admin && DISABLE_NEXT_APP=true node ./../../bin/create-graphql-schema.js",
-    "maketypes": "yarn maketypes:schema && yarn maketypes:legacy && yarn maketypes:nextgen",
+    "maketypes": "mkdirp dist/admin && DISABLE_NEXT_APP=true node ./../../bin/create-graphql-schema.js && graphql-codegen -c codegen-schema.ts",
+    "generate:client-utils": "graphql-codegen -c codegen-client-utils.ts",
     "createservice": "node ./../../bin/createservice",
     "createschema": "node ./../../bin/createschema",
     "lint-schema": "DISABLE_NEXT_APP=true node ./../../bin/lint-schema.js",
@@ -31,6 +29,7 @@
     "@ant-design/colors": "^6.0.0",
     "@ant-design/icons": "^4.7.0",
     "@apollo/client": "^3.11.8",
+    "@condo/codegen": "link:./codegen",
     "@condo/domains": "link:./domains",
     "@emotion/babel-plugin": "^11.10.0",
     "@emotion/core": "^10.1.1",
@@ -187,10 +186,9 @@
   "devDependencies": {
     "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/import-types-preset": "^3.0.0",
-    "@graphql-codegen/near-operation-file-preset": "^3.0.0",
     "@graphql-codegen/typescript": "^4.0.9",
     "@graphql-codegen/typescript-operations": "^4.2.3",
-    "@graphql-codegen/typescript-react-apollo": "^4.3.2",
+    "@graphql-codegen/typescript-react-apollo": "4.0.0",
     "@testing-library/react": "^11.2.5",
     "@types/echarts": "4.9.16",
     "@types/jest": "^29.4.0",
